# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "Documentation interactive de l'API MusicBox (Artistes, Albums, Chansons)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Lister tous les albums",
                "description": "Retourne la liste paginée des albums",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Thriller"
                                            },
                                            "artist_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "release_year": {
                                                "type": "integer",
                                                "example": 1982
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Créer un nouvel album",
                "description": "Ajoute un album dans la base de données",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id",
                                    "release_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Thriller"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "release_year": {
                                        "type": "integer",
                                        "example": 1982
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Thriller"
                                        },
                                        "artist_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "release_year": {
                                            "type": "integer",
                                            "example": 1982
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Mettre à jour un album",
                "description": "Met à jour les informations d’un album existant",
                "operationId": "7ce4b18f8619c4569c95160d86e23ff3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Bad"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "release_year": {
                                        "type": "integer",
                                        "example": 1987
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Bad"
                                        },
                                        "artist_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "release_year": {
                                            "type": "integer",
                                            "example": 1987
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Supprimer un album",
                "description": "Supprime un album de la base de données",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Album supprimé avec succès"
                    },
                    "404": {
                        "description": "Album non trouvé"
                    }
                }
            }
        },
...
    },
...
}